Updated GraphNode and GraphWidget mouse logic
GraphWidget caused errors when left-clicking on mouse due to the
nearestCircle function call with no circles. This was fixed with an if
statement checking number of circles. This fix should be later
implemented into nearestCircle for better use.

Due to issues with edge connections, GraphNode had to use pointers
instead of objects to store referencing. Some changes were made, but
function inputs/outputs were mostly kept the same to avoid pointer
ownership exchanges with adding new circles.

Main Changes
============
1. GraphWidget updates on mousePress event for checking number of
   circles
2. GraphNode updates with list storing pointers instead of objects.

To-Do
=====
1. Update nearestCircle to return NULL when no circles are available.
   (Possibly make NullCircleNode class?)